Create base class to help people implement tabletable which does own symbol handling
